feat(provenance): open the full listing in a modal, not a 400px column
A 235-row grouped table does not fit a side panel. The drawer keeps the
highlighted settings inline, where a key/value table still reads at
panel width, and "All N parameters" now opens the ingestion report's own
component in a modal — the same search, grouping and unset filter, one
implementation rather than a cramped second copy.
